fix: fixed examples bugs
